LRAC-12185 Adapt Properties to be used in pages and wizard modes
