+++ modules/apps/analytics/analytics-settings-web/package.json
@@ -1,9 +1,7 @@
 {
 	"dependencies": {
+		"@types/warning": "^3.0.0",
 		"frontend-js-web": "*"
-	},
-	"devDependencies": {
-		"@types/warning": "^3.0.0"
 	},
 	"main": "js/index.js",
 	"name": "@liferay/analytics-settings-web",
