+++ modules/apps/analytics/analytics-settings-web/src/main/resources/META-INF/resources/js/pages/default/ConnectPage.tsx
@@ -13,18 +13,11 @@
  */
 
 import React from 'react';
-export declare type TDataSource = {
-	dataSourceId: string;
-	siteIds: Array<number>;
-};
-export declare type TProperty = {
-	channelId: string;
-	commerceSyncEnabled?: boolean;
-	dataSources: Array<TDataSource>;
-	name: string;
-};
-interface IPropertiesTable {
-	properties: Array<TProperty>;
-}
-declare const PropertiesTable: React.FC<IPropertiesTable>;
-export default PropertiesTable;
+
+import Connect from '../../components/Connect';
+
+const ConnectPage = () => (
+	<Connect title={Liferay.Language.get('workspace-connection')} />
+);
+
+export default ConnectPage;
