+++ modules/apps/analytics/analytics-settings-web/src/main/resources/META-INF/resources/js/utils/request.ts
@@ -62,16 +62,19 @@ function request(path: string, config: RequestInit) {
 
 		if (response.status === Status.Forbidden) {
 			window.location.reload();
-		} else if (response.status === Status.NoContent) {
+		}
+		else if (response.status === Status.NoContent) {
 			return status;
-		} else if (response.status >= Status.MultipleChoices || status) {
+		}
+		else if (response.status >= Status.MultipleChoices || status) {
 			Liferay.Util.openToast({
 				message: ERROR_MESSAGE,
 				type: 'danger',
 			});
 
 			return status;
-		} else if (config.method === 'GET') {
+		}
+		else if (config.method === 'GET') {
 			return response.json();
 		}
