+++ modules/apps/analytics/analytics-settings-web/src/main/resources/META-INF/resources/js/utils/useFecthData.ts
@@ -38,10 +38,12 @@ function useFetchData(
 			try {
 				if (response.error) {
 					throw response.error;
-				} else {
+				}
+				else {
 					setData(response);
 				}
-			} catch (error) {
+			}
+			catch (error) {
 				console.error(error);
 
 				setError(true);
