+++ modules/apps/analytics/analytics-settings-web/types/src/main/resources/META-INF/resources/js/pages/default/ConnectPage.d.ts
@@ -12,19 +12,5 @@
  * details.
  */
 
-import React from 'react';
-export declare type TDataSource = {
-	dataSourceId: string;
-	siteIds: Array<number>;
-};
-export declare type TProperty = {
-	channelId: string;
-	commerceSyncEnabled?: boolean;
-	dataSources: Array<TDataSource>;
-	name: string;
-};
-interface IPropertiesTable {
-	properties: Array<TProperty>;
-}
-declare const PropertiesTable: React.FC<IPropertiesTable>;
-export default PropertiesTable;
+declare const ConnectPage: () => JSX.Element;
+export default ConnectPage;
