+++ modules/apps/analytics/analytics-settings-web/types/src/main/resources/META-INF/resources/js/pages/default/PropertiesPage.d.ts
@@ -13,18 +13,5 @@
  */
 
 import React from 'react';
-export declare type TDataSource = {
-	dataSourceId: string;
-	siteIds: Array<number>;
-};
-export declare type TProperty = {
-	channelId: string;
-	commerceSyncEnabled?: boolean;
-	dataSources: Array<TDataSource>;
-	name: string;
-};
-interface IPropertiesTable {
-	properties: Array<TProperty>;
-}
-declare const PropertiesTable: React.FC<IPropertiesTable>;
-export default PropertiesTable;
+declare const PropertiesPage: React.FC;
+export default PropertiesPage;
