+++ modules/apps/analytics/analytics-settings-web/types/src/main/resources/META-INF/resources/js/utils/api.d.ts
@@ -16,3 +16,5 @@ export declare function fetchConnection(token: string): Promise<Response>;
 export declare function deleteConnection(): Promise<Response>;
 export declare function fetchProperties(): Promise<any>;
 export declare function createProperty(name: string): Promise<Response>;
+export declare function fetchChannels(): Promise<any>;
+export declare function fetchSites(): Promise<any>;
